@@ -1,23 +1,15 @@
 package com.liveperson.mobilemessagingexercise;
 
 import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.EditText;
-
 
 
 /******************************************************************************
- * Class for the activity associated with the application Welcome screen
+ * Class for the activity associated with the application Welcome screen.
  * NOTE: This class also provides the listener for click events on the screen
- *****************************************************************************/
-public class WelcomeActivity extends Activity implements View.OnClickListener {
-    private static final String TAG = WelcomeActivity.class.getSimpleName();
+ ****************************************************************************implemen*/
+public class WelcomeActivity extends Activity {
 
     /**
      * Android callback invoked as the activity is created
@@ -31,29 +23,9 @@ public class WelcomeActivity extends Activity implements View.OnClickListener {
     }
 
     /**
-     * Android callback invoked as the activity is re-started by a new intent
-     * @param intent the intent associated with the restart action
-     */
-    @Override
-    protected void onNewIntent(Intent intent) {
-        super.onNewIntent(intent);
-        processLePushMessage(intent);
-    }
-
-    /**
-     * Android callback invoked as the activity is resumed
-     */
-    @Override
-    protected void onResume() {
-        super.onResume();
-        //Load saved data into the controls on this screen
-
-    }
-
-    /**
-     * Android callback invoked as the options menu is created
+     * Android callback invoked as the options menu is created.
      * @param menu the options menu in the toolbar
-     * @returns true, if the menu is to be displayed, and false otherwise
+     * @return true, if the menu is to be displayed, and false otherwise
      */
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -62,22 +34,4 @@ public class WelcomeActivity extends Activity implements View.OnClickListener {
         //Ensure the menu is displayed
         return true;
     }
-
-
-    /**
-     * Handle click events for controls on the Welcome screen
-     * @param view the control on which the event occurred
-     */
-    public void onClick(View view) {
-
-    }
-
-    /**
-     * Process a creation or restart triggered by a push message
-     * @param intent the intent associated with the push message
-     */
-    private void processLePushMessage(Intent intent) {
-
-    }
-
 }
